Fix signature check bypass
Signature check could be bypassed if `msg->len` is just right to overflow
the calculation of len.  Since the len variable is an uint16_t, it could wrap
to 0, which disregards the contents of the message for signature checking
purposes.

Fix by declaring len as an uint32_t and returning false if it exceeds
an UINT16_MAX.

Found via code inspection.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>