common.xml: add gripper hold action
Cherry-picked combination of mavlink/mavlink commits c748e54 and aa20540.

* common.xml: add gripper hold action

Grippers are not always moving, and it can be beneficial to be able to maintain the current position/state to avoid damaging the cargo and/or the device being controlled.

* common.xml: fix MAV_CMD_DO_GRIPPER min spec

Value "0" has a documented meaning, so it should be included in the values considered valid.

---------

Co-authored-by: Hamish Willee <hamishwillee@gmail.com>